Layout changes and fixes.

+++ src/components/leagues/leagues.component.js
@@ -18,21 +18,21 @@ export default class LeaguesComponent extends Component {
 
   async loadLeagues() {
     const leagues = await LeagueService.getUserLeagues()
-    this.setState({ leagues, open: false })
+    this.setState({ leagues })
   }
 
-  show = () => this.setState({ open: true })
-  handleDeleteConfirm = async (leagueId) => {
-    await LeagueService.delete(leagueId)
-    this.loadLeagues()
+  handleDelete = async (leagueId) => {
+    if (window.confirm('Opravdu smazat?')) {
+      await LeagueService.delete(leagueId)
+      this.loadLeagues()
+    }
   }
-  handleDeleteCancel = () => this.setState({ open: false })
 
   render() {
     return (
-      <div>
+      <div style={{backgroundColor: 'white'}}>
         <Header as="h1">Ligy</Header>
-        <Link to="/leagues/form/new">
+        <Link to="/league/form/new">
           <Button primary>
             Přidat ligu
           </Button>
@@ -45,29 +45,8 @@ export default class LeaguesComponent extends Component {
                 <Card.Header>
                   <Link to={`/leagues/${league.id}/matches`}>{league.name}</Link>
                 </Card.Header>
-                <Link to={`/dashboard/${league.id}`} style={{marginRight: '5px'}}>Sázky</Link>
-                <Link to={`/leagues/form/${league.id}`} style={{marginRight: '5px'}}>Upravit</Link>
-                <a href="#" onClick={this.show} style={{marginRight: '5px'}}>Smazat</a>
-                {/*<Confirm
-                  open={this.state.open}
-                  content='Opravdu smazat?'
-                  onCancel={this.handleDeleteCancel}
-                  onConfirm={() => { this.handleDeleteConfirm(league.id) }}
-                />*/}
-                <Modal size='small' open={this.state.open} onClose={this.handleDeleteCancel}>
-                  <Modal.Header>
-                    Smazat {league.name} ?
-                  </Modal.Header>
-                  <Modal.Content>
-                    <p>Chcete opravdu smazat ligu "{league.name}" ?</p>
-                  </Modal.Content>
-                  <Modal.Actions>
-                    <Button negative onClick={this.handleDeleteCancel}>
-                      Ne
-                    </Button>
-                    <Button positive onClick={() => this.handleDeleteConfirm(league.id)} icon='checkmark' labelPosition='right' content='Ano'/>
-                  </Modal.Actions>
-                </Modal>
+                <Link to={`/league/form/${league.id}`} style={{marginRight: '5px'}}>Upravit</Link>
+                <a href="#" onClick={() => this.handleDelete(league.id)} style={{marginRight: '5px'}}>Smazat</a>
               </Card.Content>
             </Card>
         ))}
